ToC + anchors + Question Combining + fixes
Added Table of Content.

fixed "Why is the like button not visible to me" as it wont be visible in 4k/packs as well

Added more description to What is the meaning of "YOU HAVE INSUFFICIENT PRIVILEGES TO REPLY HERE."? as its only certain sections

Removed I just created a thread/post but can't see it posted. and combined. In reality it's the same question

Combined Pearls Question. 

Added GOTO/Anchors

Fixed Encoder Usergroup Question